+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # CHANGELOG
 
-## unreleased [2023-02-18]
-- add tag to question
+## unreleased [2023-02-19]
+- move operations to drawer
+
+## v1.0.8 [2023-02-18]
+- add tag in app
+- attach tag to question in app
 
 ## v1.0.7 [2023-02-16]
 - add option to delete sample questions